requirements.txt
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-simple-history==3.10.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
whitenoise==6.11.0
